layers: New wsi_validation for CoreChecks
